+++ ui/src/app/admin/_components/admin-table/admin-table.tsx
@@ -6,22 +6,22 @@ import {
     getCoreRowModel,
     getPaginationRowModel,
     getFilteredRowModel,
-    getSortedRowModel
+    getSortedRowModel, SortingState,
 } from '@tanstack/react-table';
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table';
-import AdminTableColumns from '@/components/table/admin-table/table-element/admin-table-columns';
+import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow} from '@/components/ui/table';
+import AdminTableColumns from '@/app/admin/_components/admin-table/table-element/admin-table-columns';
 import PaginationBar from '@/components/table/table-element/pagination-bar';
 import GlobalFilter from '@/components/table/table-element/global-filter';
 import SortArrow from '@/components/table/table-element/sort-arrow';
 import { useState } from 'react';
 import { GroupingGroupMembers } from '@/lib/types';
 import dynamic from 'next/dynamic';
-import AdminTableSkeleton from '@/components/table/admin-table/admin-table-skeleton';
+import AdminTableSkeleton from '@/app/admin/_components/admin-table/admin-table-skeleton';
 const pageSize = parseInt(process.env.NEXT_PUBLIC_PAGE_SIZE as string);
 
 const AdminTable = ({ groupingGroupMembers }: { groupingGroupMembers: GroupingGroupMembers }) => {
     const [globalFilter, setGlobalFilter] = useState('');
-    const [sorting, setSorting] = useState([]);
+    const [sorting, setSorting] = useState<SortingState>([]);
 
     const table = useReactTable({
         columns: AdminTableColumns,
